# eight/item-api/pkg/api/item.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/item.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Item"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/item": {
      "post": {
        "summary": "Создание товара",
        "operationId": "Item_CreateItem",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/itemCreateItemResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/itemCreateItemRequest"
            }
          }
        ],
        "tags": [
          "Item"
        ]
      }
    },
    "/v1/item/{id}": {
      "get": {
        "summary": "Получение информации о товаре",
        "operationId": "Item_GetItem",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/itemGetItemResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "Item"
        ]
      }
    }
  },
  "definitions": {
    "itemCreateItemRequest": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "title": "Название"
        },
        "description": {
          "type": "string",
          "title": "Описание"
        }
      }
    },
    "itemCreateItemResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "itemGetItemResponse": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "title": "Название"
        },
        "description": {
          "type": "string",
          "title": "Описание"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}